KiCad project for High Voltage Power supply
This is an alpha version of the high-voltage power supply to drive corona in the Electrostatic Precipitator, done with KiCad. The expected input voltage will be 18VDC. The step-up transformer isn't selected yet, and the number of stages in the charge pump to reach 5kV are tentative. This version uses a 555-timer to produce a square-wave, which is then sent through a low-pass filter and then to a step-up transformer.